Suppresion des abstracts repository pour les classes qui présentaient des différences dans leur schéma
